% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryInput.R
\name{directoryInput}
\alias{directoryInput}
\title{Directory Selection Control}
\usage{
directoryInput(inputId, label, value)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value}

\item{label}{Display label for the control, or NULL for no label}

\item{value}{A directory path to set}
}
\value{
A directory input control that can be added to a UI definition.
}
\description{
Create a directory selection control to select a directory on the server
}
