% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{displayTable}
\alias{displayTable}
\title{Display dataframe in table using DT package}
\usage{
displayTable(df, session)
}
\arguments{
\item{df}{a dataframe}

\item{session}{a session}
}
\value{
A table
}
\description{
Display dataframe in table using DT package
}
\examples{

\dontrun{
session <- NULL
L3 <- LETTERS[1:3]
df <-data.frame(1, 1:10, sample(L3, 10, replace = TRUE))
displayTable(df, sesion)
}
}

