% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{dosimetry_CNRP}
\alias{dosimetry_CNRP}
\title{dosimetry_CNRP}
\format{A data frame}
\usage{
data(dosimetry_CNRP)
}
\description{
dosimetry_CNRP
}
\details{
A sample of dosimeter
}
\keyword{datasets}

