% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{dosimetry_CNSTN}
\alias{dosimetry_CNSTN}
\title{dosimetry_CNSTN}
\format{A data frame}
\usage{
data(dosimetry_CNSTN)
}
\description{
dosimetry_CNSTN
}
\details{
A sample of dosimeter
}
\keyword{datasets}

