% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosimetryPlot.R
\name{dosimetryPlot}
\alias{dosimetryPlot}
\title{plot dosimetry data}
\usage{
dosimetryPlot(df, X, Y, rangeDate, Split, Fill, Type, ReOrder)
}
\arguments{
\item{df}{dataframe with dosimetry data (column with Dates are necessary)}

\item{X}{x_variable}

\item{Y}{y_variable}

\item{rangeDate}{date or  date range}

\item{Split}{split dataframe by variable}

\item{Fill}{fill plots by variable}

\item{Type}{plot Type ('Bar', 'Scatter')}

\item{ReOrder}{reorder x axis by y (X,reorder(tmp[[X]],tmp[[Y]]))}
}
\value{
web page of radiant.dose Shiny App
}
\description{
plot dosimetry data
}
\details{
See \url{https://radiant.dose.github.io/docs} for documentation and tutorials
}
\examples{
\dontrun{
library(radiant.dose)
dat <- generate data......
dosimetryPlot(......)
}
}

