% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.dose.R
\name{launcher}
\alias{launcher}
\title{Create a launcher on the desktop for Windows (.bat), Mac (.command), or Linux (.sh)}
\usage{
launcher()
}
\description{
Create a launcher on the desktop for Windows (.bat), Mac (.command), or Linux (.sh)
}
\details{
On Windows/Mac/Linux a file named radiant.dose.bat/radiant.dose.command/radiant.dose.sh will be put on the desktop. Double-click the file to launch the specified Radiant app
}
\examples{
\dontrun{
radiant.dose::launcher()
}
}
\seealso{
\code{\link{win_launcher}} to create a shortcut on Windows

\code{\link{mac_launcher}} to create a shortcut on Mac

\code{\link{lin_launcher}} to create a shortcut on Linux
}
