% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.dose.R
\name{linux_launcher}
\alias{linux_launcher}
\title{Create a launcher and updater for Linux (.sh)}
\usage{
linux_launcher()
}
\description{
Create a launcher and updater for Linux (.sh)
}
\details{
On Linux a file named 'radiant.dose.sh' and one named 'update_radiant.dose.sh' will be put on the desktop. Double-click the file to launch the specified radiant.dose app or update radiant.dose to the latest version
}
\examples{
\dontrun{
radiant.dose::linux_launcher()
}

}
