% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.dose.R
\name{mac_launcher}
\alias{mac_launcher}
\title{Create a launcher and updater for Mac (.command)}
\usage{
mac_launcher()
}
\description{
Create a launcher and updater for Mac (.command)
}
\details{
On Mac a file named 'radiant.dose.command' and one named 'update_radiant.dose.command' will be put on the desktop. Double-click the file to launch radiant.dose or update to the latest version.
"
}
\examples{
\dontrun{
radiant.dose::mac_launcher()
}

}
