% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{Medical_CNSTN}
\alias{Medical_CNSTN}
\title{Medical_CNSTN}
\format{A data frame}
\usage{
data(Medical_CNSTN)
}
\description{
Medical_CNSTN
}
\details{
A sample of Medical essays
}
\keyword{datasets}
