% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.dose.R
\name{radiant.dose}
\alias{radiant.dose}
\title{Launch radiant.dose in the default browser}
\usage{
radiant.dose(navPanel.Edit=FALSE)
}
\arguments{
\item{navPanel.Edit}{choose panel option}
}
\value{
web page of bioCancer Shiny App
}
\description{
Launch radiant.dose in the default browser
}
\details{
See \url{https://Dosimetrix.github.io/docs} for documentation and tutorials
}
\examples{
\dontrun{
library(Dosimetrix)
radiant.dose(navPanel.Edit=FALSE)
}
}

