% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{responses}
\alias{responses}
\title{responses}
\format{A data frame with 3000 rows and 10 variables}
\usage{
data(responses)
}
\description{
responses
}
\details{
A sample of 3,000 from the diamonds dataset bundeled with ggplot2. Description provided in attr(diamonds,"description")
}
\keyword{datasets}

