% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Upper_threshold.R
\name{Upper_threshold}
\alias{Upper_threshold}
\title{display in table patients with excessive cumulative doses}
\usage{
Upper_threshold(dat, rangeDate,Y)
}
\arguments{
\item{dat}{dataframe with dosimetry data (column with Dates are necessary)}

\item{rangeDate}{A range of date}

\item{Y}{numeric variable (DCE, DP)}
}
\value{
table with excessive cumulative doses up to threshold (mSv)
}
\description{
display in table patients with excessive cumulative doses
}
\details{
See \url{https://radiant.dose.github.io/docs} for documentation and tutorials
}
\examples{
\dontrun{
 Upper_threshold(dat, rangeDate,Y)
}
}
