% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.dose.R
\name{win_launcher}
\alias{win_launcher}
\title{Create a launcher and updater for Windows (.bat)}
\usage{
win_launcher()
}
\description{
Create a launcher and updater for Windows (.bat)
}
\details{
On Windows a file named 'radiant.dose.bat' and one named 'update_radiant.command will be put on the desktop. Double-click the file to launch radiant.dose or update to the latest version.
}
\examples{
\dontrun{
radiant.dose::win_launcher()
}

}
