% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosi_report.R
\name{dosi_report}
\alias{dosi_report}
\title{display dosimetry report}
\usage{
dosi_report(dat)
}
\arguments{
\item{dat}{dataframe with dosimetry data (column with Dates are necessary)}
}
\value{
individual table with DCE and DP doses and sum (mSv)

Date of next sampling
}
\description{
display dosimetry report
}
\details{
See \url{https://radiant.dose.github.io/docs} for documentation and tutorials
}
\examples{
\dontrun{
}
}

