% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/radiant.dose.R
\docType{package}
\name{radiant.dose}
\alias{radiant.dose}
\alias{radiant.dose-package}
\title{radiant.dose}
\usage{
radiant.dose(navPanel.Edit=FALSE)
}
\arguments{
\item{navPanel.Edit}{choose panel option}
}
\value{
web page of bioCancer Shiny App
}
\description{
radiant.dose

Launch radiant.dose in the default browser
}
\details{
See \url{https://Dosimetrix.github.io/docs} for documentation and tutorials
}
\examples{
\dontrun{
library(radiant.dose)
radiant.dose(navPanel.Edit=FALSE)
}
}

